homework_1/requirements.txt
fastapi[all]==0.104.1
minio==7.1.17
names==0.3.0
python-dotenv==1.0.0
tqdm==4.66.1
ujson==5.8.0
